requirements.txt
# Core dependencies
neo4j==5.16.0
pandas==2.2.0
tqdm==4.66.0
pyyaml==6.0.0
numpy==1.24.3
weaviate-client==3.26.7

# Embedding and translation dependencies
sentence-transformers==2.2.2
transformers==4.36.2
accelerate==0.20.0
torch==2.0.0
huggingface-hub==0.21.4

# Tokenizer dependencies
sentencepiece==0.1.99
tokenizers==0.15.2
tiktoken==0.6.0
protobuf==3.20.3

# Optional: GPU support for Apple Silicon
# Note: Install these only if using M1/M2/M3 Mac
# torch-mps  # Available through conda: conda install pytorch -c pytorch

# Additional dependencies
pillow==10.3.0